Max/drop pk (#83)
Adds Add/Drop PK syntax

+++ go/vt/sqlparser/ast.go
@@ -1618,6 +1618,7 @@ const (
 	FulltextStr   = "fulltext"
 	SetStr        = "set"
 	TemporaryStr  = "temporary"
+	PrimaryStr    = "primary"
 )
 
 // Format formats the node.
@@ -2295,12 +2296,25 @@ func (idx *IndexSpec) Format(buf *TrackedBuffer) {
 	case "create", "CREATE":
 		buf.Myprintf("add ")
 		if idx.Type != "" {
-			buf.Myprintf("%s ", idx.Type)
+			if idx.Type == PrimaryStr {
+				if idx.ToName.val == "" {
+					buf.Myprintf("primary key ")
+				} else {
+					buf.Myprintf("constraint %s primary key ", idx.ToName.val)
+				}
+			} else {
+				buf.Myprintf("%s ", idx.Type)
+			}
 		}
-		buf.Myprintf("index %s ", idx.ToName.val)
+
+		if idx.Type != PrimaryStr {
+			buf.Myprintf("index %s ", idx.ToName.val)
+		}
+
 		if idx.Using.val != "" {
 			buf.Myprintf("using %s ", idx.Using.val)
 		}
+
 		buf.Myprintf("(")
 		for i, col := range idx.Columns {
 			if i != 0 {
@@ -2325,7 +2339,11 @@ func (idx *IndexSpec) Format(buf *TrackedBuffer) {
 			}
 		}
 	case "drop", "DROP":
-		buf.Myprintf("drop index %s", idx.ToName.val)
+		if idx.Type == PrimaryStr {
+			buf.Myprintf("drop primary key")
+		} else {
+			buf.Myprintf("drop index %s", idx.ToName.val)
+		}
 	case "rename", "RENAME":
 		buf.Myprintf("rename index %s to %s", idx.FromName.val, idx.ToName.val)
 	}
